+++ sea-query-driver/src/sqlx_mysql.rs
@@ -1,10 +1,10 @@
-use crate::utils::Args;
+use crate::utils::{BindParamArgs, SqlxDriverArgs};
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::parse_macro_input;
 
 pub fn bind_params_sqlx_mysql_impl(input: TokenStream) -> TokenStream {
-    let args = parse_macro_input!(input as Args);
+    let args = parse_macro_input!(input as BindParamArgs);
     let query = args.query;
     let params = args.params;
 
@@ -69,12 +69,13 @@ pub fn bind_params_sqlx_mysql_impl(input: TokenStream) -> TokenStream {
                     };
                 }
                 macro_rules! bind_box {
-                    ( $v: expr ) => {
-                        match $v {
-                            Some(v) => query.bind(v.as_ref()),
-                            None => query.bind(None::<bool>),
-                        }
-                    };
+                    ( $v: expr ) => {{
+                        let v = match $v {
+                            Some(v) => Some(v.as_ref()),
+                            None => None,
+                        };
+                        query.bind(v)
+                    }};
                 }
                 query = match value {
                     Value::Bool(v) => bind!(v, bool),
@@ -106,24 +107,28 @@ pub fn bind_params_sqlx_mysql_impl(input: TokenStream) -> TokenStream {
     output.into()
 }
 
-pub fn sea_query_driver_mysql_impl() -> TokenStream {
+pub fn sea_query_driver_mysql_impl(input: TokenStream) -> TokenStream {
+    let args = parse_macro_input!(input as SqlxDriverArgs);
+    let sqlx_path = args.driver;
+    let sea_query_path = args.sea_query;
+
     let output = quote! {
         mod sea_query_driver_mysql {
-            use sqlx::{mysql::MySqlArguments, MySql};
-            use sea_query::{Value, Values};
+            use #sqlx_path::sqlx::{mysql::MySqlArguments, MySql};
+            use #sea_query_path::sea_query::{Value, Values};
 
-            type SqlxQuery<'a> = sqlx::query::Query<'a, MySql, MySqlArguments>;
-            type SqlxQueryAs<'a, T> = sqlx::query::QueryAs<'a, MySql, T, MySqlArguments>;
+            type SqlxQuery<'a> = #sqlx_path::sqlx::query::Query<'a, MySql, MySqlArguments>;
+            type SqlxQueryAs<'a, T> = #sqlx_path::sqlx::query::QueryAs<'a, MySql, T, MySqlArguments>;
 
             pub fn bind_query<'a>(query: SqlxQuery<'a>, params: &'a Values) -> SqlxQuery<'a> {
-                sea_query::bind_params_sqlx_mysql!(query, params.0)
+                #sea_query_path::sea_query::bind_params_sqlx_mysql!(query, params.0)
             }
 
             pub fn bind_query_as<'a, T>(
                 query: SqlxQueryAs<'a, T>,
                 params: &'a Values,
             ) -> SqlxQueryAs<'a, T> {
-                sea_query::bind_params_sqlx_mysql!(query, params.0)
+                #sea_query_path::sea_query::bind_params_sqlx_mysql!(query, params.0)
             }
         }
     };